recallgpt/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers==3.0.0
huggingface-hub==0.20.0
requests==2.31.0
python-dotenv==1.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
torch>=2.0.0
numpy